+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 1.2.0
 
-* Flutter Plugin for sharing contents to social media.You can use it share to Mail, Phone, WhatsApp And System Share UI.
+* Open local file (only Android, iOS).
 
 ## 1.1.0